+++ lib/src/registry/type_registry.dart
@@ -5,7 +5,5 @@ part of hive;
 /// TypeIds have to be unique and must not change.
 abstract class TypeRegistry {
   /// Register a [TypeAdapter] to announce it to Hive.
-  ///
-  /// This is a necessary step before custom adapter can be used by Hive.
-  void registerAdapter<T>(TypeAdapter<T> adapter, [int typeId]);
+  void registerAdapter<T>(TypeAdapter<T> adapter);
 }